rembg-service/requirements.txt
rembg[cpu]==2.0.59
flask==3.1.0
gunicorn==23.0.0
Pillow==11.0.0
onnxruntime==1.19.2